@@ -28,20 +28,20 @@
     "docs": "https://docs.typo3.org/p/yoast-seo-for-typo3/yoast_seo/main/en-us/"
   },
   "require": {
-    "typo3/cms-core": "^10.4 || ^11.5 || ^12.4",
-    "typo3/cms-backend": "^10.4 || ^11.5 || ^12.4",
-    "typo3/cms-extbase": "^10.4 || ^11.5 || ^12.4",
-    "typo3/cms-fluid": "^10.4 || ^11.5 || ^12.4",
-    "typo3/cms-frontend": "^10.4 || ^11.5 || ^12.4",
-    "typo3/cms-install": "^10.4 || ^11.5 || ^12.4",
-    "typo3/cms-seo": "^10.4 || ^11.5 || ^12.4",
+    "typo3/cms-core": "^11.5 || ^12.4 || ^13.0",
+    "typo3/cms-backend": "^11.5 || ^12.4 || ^13.0",
+    "typo3/cms-extbase": "^11.5 || ^12.4 || ^13.0",
+    "typo3/cms-fluid": "^11.5 || ^12.4 || ^13.0",
+    "typo3/cms-frontend": "^11.5 || ^12.4 || ^13.0",
+    "typo3/cms-install": "^11.5 || ^12.4 || ^13.0",
+    "typo3/cms-seo": "^11.5 || ^12.4 || ^13.0",
     "ext-curl": "*",
     "ext-json": "*",
-    "php": "^7.4 || ^8.0"
+    "php": "^8.0"
   },
   "require-dev": {
     "friendsofphp/php-cs-fixer": "^3.0",
-    "overtrue/phplint": "^3.0 || ^4.0 || ^5.0",
+    "overtrue/phplint": ">=5.5 <10.0",
     "phpstan/phpstan": "^1.9",
     "phpstan/extension-installer": "^1.0",
     "typo3/tailor": "^1.1",
@@ -59,10 +59,6 @@
     }
   },
   "extra": {
-    "branch-alias": {
-      "dev-main": "9.x-dev",
-      "dev-master": "9.x-dev"
-    },
     "typo3/cms": {
       "extension-key": "yoast_seo"
     }
