@@ -11,7 +11,6 @@ use TYPO3\CMS\Core\Http\HtmlResponse;
 use TYPO3\CMS\Core\Http\JsonResponse;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use YoastSeoForTypo3\YoastSeo\Service\CrawlerService;
-use YoastSeoForTypo3\YoastSeo\Service\DbalService;
 use YoastSeoForTypo3\YoastSeo\Service\LinkingSuggestionsService;
 use YoastSeoForTypo3\YoastSeo\Service\PreviewService;
 use YoastSeoForTypo3\YoastSeo\Service\ProminentWordsService;
@@ -19,24 +18,13 @@ use YoastSeoForTypo3\YoastSeo\Service\UrlService;
 
 class AjaxController
 {
-    protected PreviewService $previewService;
-    protected UrlService $urlService;
-    protected ProminentWordsService $prominentWordsService;
-    protected LinkingSuggestionsService $linkingSuggestionsService;
-    protected CrawlerService $crawlerService;
-
     public function __construct(
-        PreviewService $previewService,
-        UrlService $urlService,
-        ProminentWordsService $prominentWordsService,
-        LinkingSuggestionsService $linkingSuggestionsService,
-        CrawlerService $crawlerService
+        protected PreviewService $previewService,
+        protected UrlService $urlService,
+        protected ProminentWordsService $prominentWordsService,
+        protected LinkingSuggestionsService $linkingSuggestionsService,
+        protected CrawlerService $crawlerService
     ) {
-        $this->previewService = $previewService;
-        $this->urlService = $urlService;
-        $this->prominentWordsService = $prominentWordsService;
-        $this->linkingSuggestionsService = $linkingSuggestionsService;
-        $this->crawlerService = $crawlerService;
     }
 
     public function previewAction(
@@ -45,7 +33,7 @@ class AjaxController
         $queryParams = $request->getQueryParams();
 
         if (!isset($queryParams['pageId'], $queryParams['languageId'], $queryParams['additionalGetVars'])) {
-            $json = json_decode($request->getBody()->getContents(), true);
+            $json = $this->getJsonData($request);
             if (isset($json['pageId'], $json['languageId'], $json['additionalGetVars'])) {
                 $queryParams = $json;
             } else {
@@ -68,9 +56,7 @@ class AjaxController
     public function saveScoresAction(
         ServerRequestInterface $request
     ): ResponseInterface {
-        $json = $request->getBody()->getContents();
-        $data = json_decode($json, true);
-
+        $data = $this->getJsonData($request);
         if (!empty($data['table']) && !empty($data['uid'])) {
             $this->saveScores($data);
         }
@@ -80,14 +66,12 @@ class AjaxController
     protected function saveScores(array $data): void
     {
         $connection = GeneralUtility::makeInstance(ConnectionPool::class)->getConnectionForTable($data['table']);
-        $row = GeneralUtility::makeInstance(DbalService::class)->getSingleResult(
-            $connection->select(['*'], $data['table'], ['uid' => (int)$data['uid']], [], [], 1)
-        );
+        $row = $connection->select(['*'], $data['table'], ['uid' => (int)$data['uid']], [], [], 1)->fetchAssociative();
 
         if ($row !== false && isset($row['tx_yoastseo_score_readability'], $row['tx_yoastseo_score_seo'])) {
             $connection->update($data['table'], [
                 'tx_yoastseo_score_readability' => (string)$data['readabilityScore'],
-                'tx_yoastseo_score_seo' => (string)$data['seoScore']
+                'tx_yoastseo_score_seo' => (string)$data['seoScore'],
             ], ['uid' => (int)$data['uid']]);
         }
     }
@@ -95,16 +79,15 @@ class AjaxController
     public function promimentWordsAction(
         ServerRequestInterface $request
     ): ResponseInterface {
-        $body = $request->getBody()->getContents();
-        $data = json_decode($body);
+        $data = $this->getJsonData($request);
 
-        if (isset($data->words, $data->uid)) {
+        if (isset($data['words'], $data['uid'])) {
             $this->prominentWordsService->saveProminentWords(
-                (int)$data->uid,
-                isset($data->pid) ? (int)$data->pid : null,
-                $data->table ?? 'pages',
-                (int)($data->languageId ?? 0),
-                (array)$data->words
+                (int)$data['uid'],
+                isset($data['pid']) ? (int)$data['pid'] : null,
+                $data['table'] ?? 'pages',
+                (int)($data['languageId'] ?? 0),
+                (array)$data['words']
             );
         }
 
@@ -114,13 +97,12 @@ class AjaxController
     public function internalLinkingSuggestionsAction(
         ServerRequestInterface $request
     ): ResponseInterface {
-        $body = $request->getBody()->getContents();
-        $data = json_decode($body);
+        $data = $this->getJsonData($request);
 
-        $words = $data->words ?? [];
-        $excludedPageId = (int)($data->excludedPage ?? 0);
-        $languageId = (int)($data->languageId ?? 0);
-        $content = (string)($data->content ?? '');
+        $words = $data['words'] ?? [];
+        $excludedPageId = (int)($data['excludedPage'] ?? 0);
+        $languageId = (int)($data['languageId'] ?? 0);
+        $content = (string)($data['content'] ?? '');
 
         $links = $this->linkingSuggestionsService->getLinkingSuggestions(
             $words,
@@ -133,7 +115,7 @@ class AjaxController
             'OK',
             'links' => $links,
             'excludedPage' => $excludedPageId,
-            'languageId' => $languageId
+            'languageId' => $languageId,
         ]);
     }
 
@@ -144,11 +126,11 @@ class AjaxController
         $amount = $this->crawlerService->getAmountOfPages($crawlerData['site'], $crawlerData['language']);
         if ($amount > 0) {
             return new JsonResponse([
-                'amount' => $amount
+                'amount' => $amount,
             ]);
         }
         return new JsonResponse([
-            'error' => 'No pages found to analyse'
+            'error' => 'No pages found to analyse',
         ]);
     }
 
@@ -169,14 +151,20 @@ class AjaxController
 
     protected function getCrawlerRequestData(ServerRequestInterface $request): array
     {
-        $crawlerData = json_decode($request->getBody()->getContents(), true);
+        $crawlerData = $this->getJsonData($request);
         if (!isset($crawlerData['site'], $crawlerData['language'])) {
             die(json_encode(['error' => 'No site and language provided by request']));
         }
         return [
             'site' => (int)$crawlerData['site'],
             'language' => (int)$crawlerData['language'],
-            'offset' => (int)($crawlerData['offset'] ?? 0)
+            'offset' => (int)($crawlerData['offset'] ?? 0),
         ];
+    }
+
+    protected function getJsonData(ServerRequestInterface $request): array
+    {
+        $body = $request->getBody()->getContents();
+        return json_decode($body, true);
     }
 }
