@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace YoastSeoForTypo3\YoastSeo\Service;
 
-use TYPO3\CMS\Core\Cache\CacheManager;
+use TYPO3\CMS\Core\Cache\Frontend\FrontendInterface;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Registry;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
@@ -17,21 +17,10 @@ class CrawlerService
     protected const REGISTRY_KEY = 'crawler-%d-%d';
     protected const INDEX_CHUNK = 50;
 
-    /**
-     * @var \TYPO3\CMS\Core\Cache\Frontend\FrontendInterface
-     */
-    protected $cache;
-
-    /**
-     * @var \TYPO3\CMS\Core\Registry
-     */
-    protected $registry;
-
-    public function __construct()
-    {
-        $this->cache = GeneralUtility::makeInstance(CacheManager::class)->getCache('pages');
-        $this->registry = GeneralUtility::makeInstance(Registry::class);
-    }
+    public function __construct(
+        protected FrontendInterface $cache,
+        protected Registry $registry
+    ) {}
 
     public function getAmountOfPages(int $site, int $languageId): int
     {
@@ -51,7 +40,7 @@ class CrawlerService
             'pages' => array_splice($pagesToIndex, $currentOffset, self::INDEX_CHUNK),
             'current' => $currentOffset,
             'nextOffset' => $currentOffset + self::INDEX_CHUNK,
-            'total' => $total
+            'total' => $total,
         ];
     }
 
@@ -71,7 +60,7 @@ class CrawlerService
             sprintf(self::REGISTRY_KEY, $site, $languageId),
             [
                 'offset' => $offset,
-                'total' => $total
+                'total' => $total,
             ]
         );
     }
@@ -106,11 +95,11 @@ class CrawlerService
                         $queryBuilder->expr()->in(
                             'uid',
                             $treeChunk
-                        )
+                        ),
                     ];
                 }
 
-                $statement = $queryBuilder->select($select)
+                $pages = $queryBuilder->select($select)
                     ->from('pages')
                     ->where(
                         $queryBuilder->expr()->in(
@@ -118,8 +107,7 @@ class CrawlerService
                             YoastUtility::getAllowedDoktypes()
                         ),
                         ...$constraints
-                    )->execute();
-                $pages = GeneralUtility::makeInstance(DbalService::class)->getAllResults($statement);
+                    )->executeQuery()->fetchAllAssociative();
                 $pagesToIndex = array_merge($pagesToIndex, array_column($pages, $select));
             }
         }
