@@ -8,7 +8,6 @@ use Doctrine\DBAL\Result;
 use TYPO3\CMS\Core\Database\Connection;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use YoastSeoForTypo3\YoastSeo\Service\DbalService;
 use YoastSeoForTypo3\YoastSeo\Utility\YoastUtility;
 
 class OrphanedContentDataProvider extends AbstractOverviewDataProvider
@@ -40,7 +39,7 @@ class OrphanedContentDataProvider extends AbstractOverviewDataProvider
         return $qb->select('*')
             ->from('pages')
             ->where(...$constraints)
-            ->execute();
+            ->executeQuery();
     }
 
     protected function setReferencedPages(): void
@@ -58,15 +57,14 @@ class OrphanedContentDataProvider extends AbstractOverviewDataProvider
             )
         ];
 
-        $statement = $qb->select('ref_uid')
+        $references = $qb->select('ref_uid')
             ->from('sys_refindex')
             ->where(...$constraints)
             ->groupBy('ref_uid')
-            ->execute();
+            ->executeQuery()
+            ->fetchAllAssociative();
 
-        $refs = GeneralUtility::makeInstance(DbalService::class)->getAllResults($statement);
-
-        foreach ($refs as $ref) {
+        foreach ($references as $ref) {
             $this->referencedPages[] = $ref['ref_uid'];
         }
     }
