@@ -8,13 +8,7 @@ use YoastSeoForTypo3\YoastSeo\Record\Builder\AbstractBuilder;
 
 abstract class AbstractListener
 {
-    /**
-     * @var \YoastSeoForTypo3\YoastSeo\Record\Builder\AbstractBuilder
-     */
-    protected AbstractBuilder $builder;
-
-    public function __construct(AbstractBuilder $builder)
-    {
-        $this->builder = $builder;
-    }
+    public function __construct(
+        protected AbstractBuilder $builder
+    ) {}
 }
