@@ -8,8 +8,10 @@ use TYPO3\CMS\Core\Authentication\BackendUserAuthentication;
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Type\Bitmask\Permission;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
-use YoastSeoForTypo3\YoastSeo\Service\DbalService;
 
+/**
+ * @deprecated Will be removed once TYPO3 v11 support is dropped
+ */
 class PagesWithoutDescriptionDataProvider implements PageProviderInterface
 {
     private array $excludedDoktypes;
@@ -28,7 +30,7 @@ class PagesWithoutDescriptionDataProvider implements PageProviderInterface
         $constraints = [
             $queryBuilder->expr()->notIn('doktype', $this->excludedDoktypes),
             $queryBuilder->expr()->eq('no_index', 0),
-            $queryBuilder->expr()->orX(
+            $queryBuilder->expr()->or(
                 $queryBuilder->expr()->eq('description', $queryBuilder->createNamedParameter('')),
                 $queryBuilder->expr()->isNull('description')
             ),
@@ -39,18 +41,18 @@ class PagesWithoutDescriptionDataProvider implements PageProviderInterface
         $iterator = 0;
 
         while ($counter < $this->limit) {
-            $statement = $queryBuilder
+            $row = $queryBuilder
                 ->select('*')
                 ->from('pages')
                 ->where(...$constraints)
                 ->orderBy('tstamp', 'DESC')
                 ->setFirstResult($iterator)
                 ->setMaxResults(1)
-                ->execute();
-            $row = GeneralUtility::makeInstance(DbalService::class)->getSingleResult($statement);
+                ->executeQuery()
+                ->fetchAssociative();
 
             if ($row === false) {
-                // Likely less pages than the limit, prevent infinite loop
+                // Likely fewer pages than the limit, prevent infinite loop
                 break;
             }
 
