@@ -1,9 +1,11 @@
 <?php
 
+use YoastSeoForTypo3\YoastSeo\Middleware\PageRequestMiddleware;
+
 return [
     'frontend' => [
         'yoast-seo-page-request' => [
-            'target' => \YoastSeoForTypo3\YoastSeo\Middleware\PageRequestMiddleware::class,
+            'target' => PageRequestMiddleware::class,
             'before' => [
                 'typo3/cms-frontend/tsfe'
             ],
