@@ -12,8 +12,8 @@ $EM_CONF[$_EXTKEY] = [
     'version' => '9.0.3',
     'constraints' => [
         'depends' => [
-            'typo3' => '10.4.0-12.4.99',
-            'seo' => '10.4.0-12.4.99',
+            'typo3' => '11.5.0-13.4.99',
+            'seo' => '11.5.0-13.4.99',
         ]
     ],
     'autoload' => [
